Move some code to cpp files
